[Refactor] Unify mtp_proposer and eagle_proposer
